chmod_at test: widen the Darwin Stat_t mode before masking
syscall.Stat_t.Mode is uint16 on Darwin, so go vet refused the helper
there and the macOS lane stopped before the corpus ran.

+++ internal/e2e/chmod_at_test.go
@@ -131,7 +131,7 @@ func chmodAtLstatMode(t *testing.T, dir, name string) uint32 {
 	if err := syscall.Lstat(filepath.Join(dir, name), &st); err != nil {
 		t.Fatalf("lstat %s: %v", name, err)
 	}
-	return st.Mode & 0o7777
+	return uint32(st.Mode) & 0o7777
 }
 
 // chmodAtSeed builds the tree chmodAtSource operates on and returns the mode
